Changeddefault in mbe_conf to be localhost for getting umask 022. Fixed platform for Mac OS X. Fixed bash_profile to not try to do hushlogin if /etc/motd does not exist.

+++ bash_profile
@@ -27,10 +27,12 @@
 #   ~/.bashrc
 # fi
 
-cmp -s $HOME/.hushlogin /etc/motd
-if [ $? != 0 ]; then
-  tee $HOME/.hushlogin < /etc/motd
-  echo -n "Press Enter to continue: " && read ans
+if [ -f "/etc/motd" ]; then
+  cmp -s $HOME/.hushlogin /etc/motd
+  if [ $? != 0 ]; then
+    tee $HOME/.hushlogin < /etc/motd
+    echo -n "Press Enter to continue: " && read ans
+  fi
 fi
 
 # Get the aliases and functions
